+++ examples/bad/tests.py
@@ -1,0 +1,6 @@
+# Check that files named "tests.py" are checked
+
+
+def test():
+    action = sorted([3, 2, 1])
+    assert action == [1, 2, 3]